Prevent a crash in the rotating cube
GKL assumes that all cubes have four sides and exactly four sides.  The code to grab the frontmostFacingViewController also assumed that but would crash if somehow there were fewer than four sides loaded.  Could be a timing issue.  This prevents that crash (but not the underlying issue that may have caused it in the first place).

TESTING:
Verify that all Discover functionality still works properly.